+++ backend/src/io/http/mod.rs
@@ -6,4 +6,5 @@ pub mod media;
 pub mod metainfo;
 pub mod page;
 mod page_type;
+pub mod state;
 pub mod r#static;