+++ backend/src/looksyk/builtinpage/mod.rs
@@ -1,3 +1,4 @@
+pub mod asset_metainfo_table;
 pub mod assets_overview;
 mod generating_page_util;
 pub mod page_not_found;