+++ backend/src/looksyk/builtinpage/page_not_found.rs
@@ -1,4 +1,4 @@
-use crate::looksyk::builder::text_token;
+use crate::looksyk::builder::text_token_str;
 use crate::looksyk::model::{BlockContent, ParsedBlock, ParsedMarkdownFile};
 
 pub fn generate_page_not_found() -> ParsedMarkdownFile {
@@ -6,7 +6,7 @@ pub fn generate_page_not_found() -> ParsedMarkdownFile {
         blocks: vec![ParsedBlock {
             content: vec![BlockContent {
                 as_text: "".to_string(),
-                as_tokens: vec![text_token(
+                as_tokens: vec![text_token_str(
                     "Page not (yet) created. Click here to write content!",
                 )],
             }],