+++ frontend/looksyk/src/app/pages/media-details/details.component.ts
@@ -1,16 +1,19 @@
-import {ChangeDetectionStrategy, Component, inject, OnInit} from '@angular/core';
-import {CommonModule} from '@angular/common';
-import {BehaviorSubject, Subject} from "rxjs";
-import {ActivatedRoute} from "@angular/router";
-import {MediaPreview, MediaService} from "../../services/media.service";
-import {MarkdownComponent} from "../components/markdown/markdown.component";
-import {DomSanitizer, SafeResourceUrl} from "@angular/platform-browser";
-import {HistoryService} from "../../services/history.service";
+import { ChangeDetectionStrategy, Component, inject, OnInit } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { BehaviorSubject, firstValueFrom, Subject } from "rxjs";
+import { ActivatedRoute } from "@angular/router";
+import { MediaPreview, MediaService } from "../../services/media.service";
+import { MarkdownComponent } from "../components/markdown/markdown.component";
+import { DomSanitizer, SafeResourceUrl } from "@angular/platform-browser";
+import { HistoryService } from "../../services/history.service";
+import { MarkdownPage } from "../model";
+import { MatIconModule } from "@angular/material/icon";
+import { MatButtonModule } from "@angular/material/button";
 
 @Component({
   selector: 'app-media-details-overview',
   standalone: true,
-  imports: [CommonModule, MarkdownComponent],
+  imports: [CommonModule, MarkdownComponent, MatIconModule, MatButtonModule],
   templateUrl: './details.component.html',
   styleUrls: ['./details.component.css'],
   changeDetection: ChangeDetectionStrategy.OnPush
@@ -26,7 +29,6 @@ export class DetailsComponent implements OnInit {
 
   private historyService = inject(HistoryService);
 
-
   public mediaInfo: Subject<MediaPreview> = new BehaviorSubject({
     properties: {
       size: "",
@@ -34,6 +36,15 @@ export class DetailsComponent implements OnInit {
     }
   });
   public mediaInfo$ = this.mediaInfo.asObservable();
+
+
+  public mediaMetadata: Subject<MarkdownPage> = new BehaviorSubject<MarkdownPage>({
+    name: "",
+    pageid: "",
+    isFavourite: false,
+    blocks: []
+  });
+  public mediaMetadata$ = this.mediaMetadata.asObservable();
 
   bypass(trustedUrl: string): SafeResourceUrl {
     return this.sanitizer.bypassSecurityTrustResourceUrl(trustedUrl.toString());
@@ -48,8 +59,23 @@ export class DetailsComponent implements OnInit {
         this.mediaService.getMediaPreviewInfo(pageName).subscribe(
           mediaPreview => this.mediaInfo.next(mediaPreview)
         );
+        this.mediaService.getMediaMetadata(pageName).subscribe(
+          metadata =>
+            this.mediaMetadata.next(metadata)
+        )
         this.historyService.pushEntry(pageName, ["/assets/", pageName]);
       });
+  }
+
+  async download() {
+    const filename = await firstValueFrom(this.pageName$)
+    const link = document.createElement('a');
+    link.setAttribute('target', '_blank');
+    link.setAttribute('href', "/assets/" + encodeURIComponent(filename));
+    link.setAttribute('download', filename);
+    document.body.appendChild(link);
+    link.click();
+    link.remove();
   }
 
 }